Organize sidebar links to a subdirectory of data
