+++ www/src/pages/features.js
@@ -6,7 +6,7 @@ import EvaluationTable from "../components/evaluation-table"
 import EvaluationCell from "../components/evaluation-cell"
 import FuturaParagraph from "../components/futura-paragraph"
 import PageWithSidebar from "../components/page-with-sidebar"
-import featuresSidebar from "../data/features-links.yaml"
+import featuresSidebar from "../data/sidebars/features-links.yaml"
 import Container from "../components/container"
 import { options, rhythm } from "../utils/typography"
 import presets, { colors } from "../utils/presets"
